debundle: split facts/mod.rs into submodules
facts/mod.rs was a ~2.8k-line module. Split it into cohesive
submodules behind the same public surface; mod.rs is now a thin
module index.

  item_views             top-level item view types
  collector              fact collection visitors
  analyze                fact analysis entry points
  statement_facts        per-statement fact derivation
  at_init_fallback       at-init unresolved-call fallback
  purity_classification  purity classification glue
  global_object          global-object reference detection
  lazy_boundary          lazy-boundary detection
  local_effect_targets   local effect-target tracking

No behavior change; all 116 debundle tests pass, clippy clean.
