+++ devinfra/js/debundle/BUILD.bazel
@@ -357,8 +357,17 @@ rust_library(
         "atomic_units.rs",
         "cross_module_purity.rs",
         "factor_assembly.rs",
+        "facts/analyze.rs",
+        "facts/at_init_fallback.rs",
+        "facts/collector.rs",
+        "facts/global_object.rs",
+        "facts/item_views.rs",
+        "facts/lazy_boundary.rs",
+        "facts/local_effect_targets.rs",
         "facts/local_effects.rs",
         "facts/mod.rs",
+        "facts/purity_classification.rs",
+        "facts/statement_facts.rs",
         "facts/wire.rs",
         "graph.rs",
         "ids.rs",
